automates test
