+++ firmware/inc/defs.h
