+++ firmware/inc/linkedlist.h
