+++ firmware/inc/settings.h
