+++ firmware/inc/usbdef.h
