+++ firmware/success
@@ -1,0 +1,3 @@
+sp 27
+Standard Keyboard Parser Test Passed
+End-to-end Standard Keyboard Test Passed
