+++ firmware/test/testargs
@@ -1,0 +1,1 @@
+break 0xffff
