+++ firmware/test/testcommon.c
@@ -16,12 +16,15 @@
 #include "keyboardled.h"
 #include "parsedescriptor.h"
 #include "ps2protocol.h"
+#include "ps2.h"
 #include "testdata.h"
 #include "preset.h"
 #include "test.h"
 #include "usbll.h"
 #include "linkedlist.h"
 #include "testcommon.h"
+#include "scancode.h"
+#include "processreport.h"
 //#define TESTVERBOSE 1
 
 /*
@@ -35,6 +38,45 @@ Things we might want to test:
 Probably gonna need seperate executables for different tests because of limited code space
  
 */
+
+__xdata bool MenuActive = 0;
+
+__xdata USB_HUB_PORT UsbDev;
+
+// mask off the number of bits
+__code uint32_t bitMasks32[] = {
+    0x0000, 
+    0x0001, 0x0003, 0x0007, 0x000f, 0x001F, 0x003F, 0x007F, 0x00FF,
+    0x01FF, 0x03FF, 0x07FF, 0x0fFF, 0x1fFF, 0x3fFF, 0x7fFF, 0xFFFF,
+    0x0001FFFF, 0x0003FFFF, 0x0007FFFF, 0x000fFFFF, 0x001FFFFF, 0x003FFFFF, 0x007FFFFF, 0x00FFFFFF,
+    0x01FFFFFF, 0x03FFFFFF, 0x07FFFFFF, 0x0fFFFFFF, 0x1fFFFFFF, 0x3fFFFFFF, 0x7fFFFFFF, 0xFFFFFFFF
+ };
+
+__at(0xF000) __code Settings DefSettings = {
+    0x54178008, //magic
+    0, //KeyboardReportMode
+    0, //MouseReportMode
+    0, //EnableAUXPS2
+    0, //Intellimouse
+    0, //XT83Keys
+    1,//GameControllerAsMouse
+    0, //SerialDebugOutput
+    0, //USBFilter
+    MODE_PS2, //KeyboardMode
+    0 //MenuRateLimit
+};
+
+__xdata volatile uint16_t SoftWatchdog = 0;
+
+void Menu_Press_Key(uint8_t key)
+{
+
+}
+
+void SetKeyboardLedStatusFromPS2(UINT8 ps2led)
+{
+
+}
 
 int putchar(int c)
 {
@@ -98,6 +140,22 @@ bool InterfaceParseDeviceDescriptors (
 
 }
 
+uint8_t * GetNextChunk(){
+    if (ports[PORT_KEY].sendBuffEnd == ports[PORT_MOUSE].sendBuffStart)
+        return NULL;
+    uint8_t * chunk = ports[PORT_KEY].sendBuff.chunky[ports[PORT_KEY].sendBuffStart];
+    ports[PORT_KEY].sendBuffStart = (ports[PORT_KEY].sendBuffStart + 1) & 0x3F;
+    return chunk;
+}
+
+uint8_t * GetNextChonk(){
+    if (ports[PORT_MOUSE].sendBuffEnd == ports[PORT_MOUSE].sendBuffStart)
+        return NULL;
+    uint8_t * chonk = ports[PORT_MOUSE].sendBuff.chonky[ports[PORT_MOUSE].sendBuffStart];
+    ports[PORT_MOUSE].sendBuffStart = (ports[PORT_MOUSE].sendBuffStart + 1) & 0x07;
+    return chonk;
+}
+
 
 void InitTest (
     USB_HUB_PORT *pUsbDevice,
@@ -134,4 +192,20 @@ __xdata HID_SEG * FindSegByStartBit(__xdata HID_REPORT * descReport, uint16_t St
 
 uint32_t rand32(void) {
     return ((uint32_t)(rand()) << 16) | (uint32_t)rand();
+}
+
+void TestSetup()
+{
+    // timer0 setup
+	TMOD = (TMOD & 0xf0) | 0x02; // mode 1 (8bit auto reload)
+	TH0 = 0x00;					 // I dunno
+
+	TR0 = 1; // start timer0
+	ET0 = 1; //enable timer0 interrupt;
+
+	EA = 1;	 // enable all interrupts
+
+    UART_Init();
+
+    printstackpointer();
 }
